--- libs/virtual-node/virtual-node-datamodel/src/main/java/net/corda/libs/virtualnode/datamodel/package-info.java
@@ -1,4 +1,0 @@
-@Export
-package net.corda.libs.virtualnode.datamodel;
-
-import org.osgi.annotation.bundle.Export;
