+++ libs/virtual-node/virtual-node-datamodel/src/main/java/net/corda/libs/virtualnode/datamodel/repository/package-info.java
@@ -1,0 +1,4 @@
+@Export
+package net.corda.libs.virtualnode.datamodel.repository;
+
+import org.osgi.annotation.bundle.Export;
