+++ libs/virtual-node/virtual-node-datamodel/src/main/kotlin/net/corda/libs/virtualnode/datamodel/repository/VirtualNodeRepositoryImpl.kt
@@ -1,6 +1,10 @@
-package net.corda.libs.virtualnode.datamodel
+package net.corda.libs.virtualnode.datamodel.repository
 
 import net.corda.libs.packaging.core.CpiIdentifier
+import net.corda.libs.virtualnode.datamodel.HoldingIdentityEntity
+import net.corda.libs.virtualnode.datamodel.VirtualNodeEntity
+import net.corda.libs.virtualnode.datamodel.VirtualNodeEntityKey
+import net.corda.libs.virtualnode.datamodel.VirtualNodeNotFoundException
 import net.corda.orm.utils.transaction
 import net.corda.v5.base.exceptions.CordaRuntimeException
 import net.corda.virtualnode.HoldingIdentity
@@ -9,14 +13,6 @@ import net.corda.virtualnode.VirtualNodeInfo
 import net.corda.virtualnode.VirtualNodeState
 import java.util.stream.Stream
 import javax.persistence.EntityManager
-
-// using an interface allows us to easily mock/test
-interface VirtualNodeRepository {
-    fun findAll(entityManager: EntityManager): Stream<VirtualNodeInfo>
-    fun find(entityManager: EntityManager, holdingIdentityShortHash: ShortHash): VirtualNodeInfo?
-    fun put(entityManager: EntityManager, holdingId: HoldingIdentity, cpiId: CpiIdentifier)
-    fun updateVirtualNodeState(entityManager: EntityManager, holdingIdentityShortHash: String, newState: VirtualNodeState): VirtualNodeInfo
-}
 
 class VirtualNodeRepositoryImpl : VirtualNodeRepository {
     /**
@@ -116,4 +112,3 @@ class VirtualNodeRepositoryImpl : VirtualNodeRepository {
             .singleOrNull()
     }
 }
-
